MCP/requirements-backup.txt
# This env packages are created based on Python 3.11.13
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
arxiv==2.1.0
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.3.0
cohere==5.20.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
faiss-cpu==1.12.0
fastapi==0.109.0
fastavro==1.12.1
feedparser==6.0.10
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib-metadata==6.11.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.5
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.8.0
networkx==3.5
nltk==3.8.1
numpy==1.24.4
oauthlib==3.3.1
onnxruntime==1.23.1
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pillow==10.4.0
plotly==6.3.1
posthog==6.7.8
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.2
pulsar-client==3.8.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.24
pydantic-settings==2.1.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.5
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.12.2
sentence-transformers==2.3.1
sentencepiece==0.2.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.35.1
streamlit==1.50.0
sympy==1.14.0
tenacity==8.2.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.66.1
transformers==4.57.1
typer==0.19.2
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.27.0
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
